+++ src/whois.h
@@ -28,5 +28,5 @@
 
 void whois_do (Netinfo *netinfo);
 void whois_stop (Netinfo *netinfo);
-void whois_foreach (Netinfo * netinfo, gchar * line, gint len, gpointer user_data);
+void whois_foreach (Netinfo * netinfo, gchar * line, gssize len, gpointer user_data);
 void whois_copy_to_clipboard (Netinfo * netinfo, gpointer user_data);
